requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
pydantic
asyncpg
sqlalchemy>=2.0
alembic